pyproject.toml
[tool.poetry]
name = "tm-tg-bot-v1"
version = "0.1.0"
description = ""
authors = ["smbrine <smbrine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
python-telegram-bot = "^21.2"
sqlalchemy = "^2.0.30"
grpcio = "^1.64.0"
grpcio-tools = "^1.64.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
redis = "^5.0.4"
matplotlib = "^3.9.0"


[tool.poetry.group.migration.dependencies]
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.2"
flake8 = "^7.0.0"
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
